Struggling with some groovy error stuff
One of the classes is an explicit generated class using the class statement, the other is a class generated from the script body based on the file name. Solutions are to change the file name or to change the class name.

maven change "local" version from 1.0.0 to 3.0.0 https://github.com/metasfresh/metasfresh/issues/2102